# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "workModes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkMode"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workmodes"
            }
          },
          {
            "name": "employmentTypes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmploymentType"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Employmenttypes"
            }
          },
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Seniority"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "minLevel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MatchLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minlevel"
            }
          },
          {
            "name": "savedOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Savedonly"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "best_match",
                "newest"
              ],
              "type": "string",
              "default": "best_match",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItem"
                  },
                  "title": "Response List Jobs Api V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/recommended": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Recommended",
        "operationId": "recommended_api_v1_jobs_recommended_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobListItem"
                  },
                  "type": "array",
                  "title": "Response Recommended Api V1 Jobs Recommended Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/import/preview": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Import Preview",
        "description": "Read a posting from a link. Nothing is saved.",
        "operationId": "import_preview_api_v1_jobs_import_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/import": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Import Job",
        "operationId": "import_job_api_v1_jobs_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Read Job",
        "operationId": "read_job_api_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/match": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Read Match",
        "operationId": "read_match_api_v1_jobs__job_id__match_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "saved",
          "preparing",
          "applied",
          "interview",
          "offer",
          "rejected"
        ],
        "title": "ApplicationStatus"
      },
...
      "DescriptionBlock": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "heading",
              "paragraph",
              "bullet"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "requirement": {
            "type": "boolean",
            "title": "Requirement",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "type",
          "text"
        ],
        "title": "DescriptionBlock"
      },
      "EmploymentType": {
        "type": "string",
        "enum": [
          "full_time",
          "part_time",
          "contract",
          "internship"
        ],
        "title": "EmploymentType"
      },
      "FactorStatus": {
        "type": "string",
        "enum": [
          "good",
          "ok",
          "poor",
          "unknown"
        ],
        "title": "FactorStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportPreview": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "hostLabel": {
            "type": "string",
            "title": "Hostlabel"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "workMode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "employmentType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "salary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Salary"
              },
              {
                "type": "null"
              }
            ]
          },
          "postedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postedat"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "items": {
              "$ref": "#/components/schemas/DescriptionBlock"
            },
            "type": "array",
            "title": "Description"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          }
        },
        "type": "object",
        "required": [
          "url",
          "hostLabel",
          "title",
          "company",
          "location",
          "workMode",
          "employmentType",
          "seniority",
          "salary",
          "postedAt",
          "summary",
          "description",
          "skills",
          "missing"
        ],
        "title": "ImportPreview",
        "description": "What was read from the link. Nothing is saved until the user confirms."
      },
      "ImportRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 8,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ImportRequest"
      },
      "JobListItem": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRead"
          },
          "match": {
            "$ref": "#/components/schemas/MatchSummary"
          },
          "applicationId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicationid"
          },
          "applicationStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job",
          "match"
        ],
        "title": "JobListItem"
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "workMode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "employmentType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "salary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Salary"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "$ref": "#/components/schemas/JobSource"
          },
          "postedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postedat"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "items": {
              "$ref": "#/components/schemas/DescriptionBlock"
            },
            "type": "array",
            "title": "Description"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "company",
          "location",
          "workMode",
          "employmentType",
          "seniority",
          "salary",
          "source",
          "postedAt",
          "summary",
          "description",
          "skills"
        ],
        "title": "JobRead"
      },
      "JobSource": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "kind": {
            "$ref": "#/components/schemas/JobSourceKind"
          }
        },
        "type": "object",
        "required": [
          "label",
          "url",
          "kind"
        ],
        "title": "JobSource"
      },
      "JobSourceKind": {
        "type": "string",
        "enum": [
          "feed",
          "imported"
        ],
        "title": "JobSourceKind"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MatchFactor": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "$ref": "#/components/schemas/FactorStatus"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "status",
          "detail"
        ],
        "title": "MatchFactor"
      },
      "MatchFactors": {
        "properties": {
          "deterministic": {
            "items": {
              "$ref": "#/components/schemas/MatchFactor"
            },
            "type": "array",
            "title": "Deterministic"
          },
          "ai": {
            "items": {
              "$ref": "#/components/schemas/MatchFactor"
            },
            "type": "array",
            "title": "Ai"
          }
        },
        "type": "object",
        "required": [
          "deterministic",
          "ai"
        ],
        "title": "MatchFactors"
      },
      "MatchGap": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "howToClose": {
            "type": "string",
            "title": "Howtoclose"
          }
        },
        "type": "object",
        "required": [
          "label",
          "howToClose"
        ],
        "title": "MatchGap"
      },
      "MatchLevel": {
        "type": "string",
        "enum": [
          "strong",
          "good",
          "partial",
          "weak"
        ],
        "title": "MatchLevel"
      },
      "MatchRead": {
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "title": "Jobid"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "level": {
            "$ref": "#/components/schemas/MatchLevel"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/MatchGap"
            },
            "type": "array",
            "title": "Gaps"
          },
          "factors": {
            "$ref": "#/components/schemas/MatchFactors"
          },
          "basedOn": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Basedon"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "score",
          "level",
          "reason",
          "strengths",
          "gaps",
          "factors",
          "basedOn",
          "generatedAt"
        ],
        "title": "MatchRead"
      },
      "MatchSummary": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "level": {
            "$ref": "#/components/schemas/MatchLevel"
          }
        },
        "type": "object",
        "required": [
          "score",
          "level"
        ],
        "title": "MatchSummary"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "RegisterRequest"
      },
...
      "Salary": {
        "properties": {
          "min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SalaryPeriod"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "min",
          "max",
          "currency",
          "period",
          "text",
          "source"
        ],
        "title": "Salary"
      },
      "SalaryPeriod": {
        "type": "string",
        "enum": [
          "year",
          "month",
          "day",
          "hour"
        ],
        "title": "SalaryPeriod"
      },
...
      "Seniority": {
        "type": "string",
        "enum": [
          "intern",
          "junior",
          "mid",
          "senior",
          "lead"
        ],
        "title": "Seniority"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPair"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkMode": {
        "type": "string",
        "enum": [
          "remote",
          "hybrid",
          "onsite"
        ],
        "title": "WorkMode"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}